MSR: Separate classiers into their own directory
Since we have a few classifiers now, we should have them in their own
directory at this point. This will also keep the testing section from
getting too long.
